+++ redisson/src/main/java/org/redisson/command/CommandAsyncExecutor.java
@@ -30,7 +30,7 @@ import org.redisson.client.codec.Codec;
 import org.redisson.client.protocol.RedisCommand;
 import org.redisson.connection.ConnectionManager;
 import org.redisson.connection.MasterSlaveEntry;
-import org.redisson.liveobject.core.RedissonObjectBuilder;
+import org.redisson.misc.RedissonObjectBuilder;
 
 /**
  *
