+++ redisson/src/main/java/org/redisson/liveobject/core/LiveObjectInterceptor.java
@@ -37,6 +37,7 @@ import net.bytebuddy.implementation.bind.annotation.FieldValue;
 import net.bytebuddy.implementation.bind.annotation.Origin;
 import net.bytebuddy.implementation.bind.annotation.RuntimeType;
 import net.bytebuddy.implementation.bind.annotation.This;
+import org.redisson.misc.RedissonObjectBuilder;
 
 /**
  *
